requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
httpx
python-dotenv
PyGithub